Create file slateUtils.ts
