+++ packages/editor/src/client/blocks/richTextBlock/slate.tsx
@@ -71,6 +71,8 @@ import {
   ListFormats
 } from './formats'
 
+import {isFormatActive, toggleFormat} from './slateUtils'
+
 const ElementTags: any = {
   A: (el: Element) => ({
     type: InlineFormat.Link,
@@ -698,52 +700,6 @@ function TableButton({icon, iconActive}: TableButtonProps) {
     </Whisper>
   )
 }
-
-function isFormatActive(editor: Editor, format: Format) {
-  if (TextFormats.includes(format)) {
-    const marks = Editor.marks(editor)
-    return marks ? marks[format] === true : false
-  }
-
-  if (BlockFormats.includes(format) || InlineFormats.includes(format)) {
-    const [match] = Editor.nodes(editor, {
-      match: node => node.type === format,
-      mode: 'all'
-    })
-
-    return !!match
-  }
-
-  return false
-}
-
-function toggleFormat(editor: Editor, format: Format) {
-  const isActive = isFormatActive(editor, format)
-  const isList = ListFormats.includes(format)
-
-  if (TextFormats.includes(format)) {
-    if (isActive) {
-      editor.removeMark(format)
-    } else {
-      editor.addMark(format, true)
-    }
-  }
-
-  if (BlockFormats.includes(format)) {
-    for (const format of ListFormats) {
-      Transforms.unwrapNodes(editor, {match: node => node.type === format, split: true})
-    }
-
-    Transforms.setNodes(editor, {
-      type: isActive ? BlockFormat.Paragraph : isList ? BlockFormat.ListItem : format
-    })
-
-    if (!isActive && isList) {
-      Transforms.wrapNodes(editor, {type: format, children: []})
-    }
-  }
-}
-
 function insertLink(editor: Editor, selection: Range | null, url: string, title?: string) {
   if (selection) {
     if (Range.isCollapsed(selection)) {
